Revert "use u32 for branch.order"
This reverts commit ce0a91c77dcb0c9936eb62e3a1f43fe3bcada835.

+++ src-tauri/src/virtual_branches/branch/reader.rs
@@ -37,7 +37,7 @@ mod tests {
 
     use super::{super::Writer, *};
 
-    static mut TEST_INDEX: u32 = 0;
+    static mut TEST_INDEX: usize = 0;
 
     fn test_branch() -> Branch {
         unsafe {
